requirements.txt
streamlit==1.28.1
streamlit-authenticator==0.2.3
pyyaml==6.0.1
bcrypt==4.1.2
streamlit-option-menu==0.3.6
streamlit-extras==0.3.6
